+++ src/signified/__init__.py
@@ -21,8 +21,22 @@ Attributes:
     NestedValue: Recursive type for arbitrarily nested reactive values.
 """
 
-from .core import Computed, Signal, Variable, as_signal, computed, deep_unref, has_value, reactive_method, unref
-from .types import HasValue, NestedValue, ReactiveValue
+from .core import (
+    Computed, 
+    Signal, 
+    Variable,
+    as_signal,
+    computed,
+    deep_unref,
+    has_value,
+    reactive_method,
+    unref,
+)
+from .types import (
+    HasValue,
+    NestedValue,
+    ReactiveValue,
+)
 
 __all__ = [
     "Variable",
